Improve permission-error-reporting and raise open-file-limit
There was a small bug, namely that when quark was executed as a normal
user, the fork would fail because setrlimit() had made it impossible
beforehand (given it only controls thread-counts). Now the setrlimit()
for thread-count is done correctly after the fork. Additionally,
given we also open at least as many files as we have threads (and each
thread needs to keep multiple fd's open at the same time), we also now
set the open-file-limit properly.

Previously, when someone tried to execute quark as non-root, they would
get the error message

   $ quark -p 5000
   quark: fork: Resource temporarily unavailable
   $

This was due to the aforementioned "bug", but even still, they would've
gotten an error message relating to a failed chroot. In either case,
it might've been a bit confusing, which is why it now shows a clear
error message on what's wrong and a possible "mitigation" (using
capabilities(7) as an alternative to setuid or root):

   $ quark -p 5000
   quark: You need to run as root or have CAP_SYS_CHROOT set
   $

CAP_SYS_CHROOT alone is not sufficient to run quark, and it will print
further errors until all permissions are met, but I won't add a separate
error handling and logic just to appease with a cumulative
error-message.

When trying to bind to a privileged port, you get

   $ quark -p 1000
   quark: You need to run as root or have CAP_NET_BIND_SERVICE set to
   bind to privileged ports
   $

instead of

   $ quark -p 1000
   quark: bind: Permission denied
   $

which is also a net-benefit.

In this context, this commit also improves the error-reporting when
someone tries 'dropping' to the root user or group by checking this
beforehand and not with a getuid() and getgid() later on.

Signed-off-by: Laslo Hunhold <dev@frign.de>