Properly set resource limits
Quark sets two rlimits, the maximum number of open file descriptors and
the maximum number of threads. I made a mistake in the calculation of
the former (forgetting about slots) and gave it a bit more headroom so
we don't run into problems. Given the number open file descriptors is
per-process, this can be considered done.

The thread-limit is a different matter, because it's per user. To work
around this, the program tries increasing the per-user-limit by the
number of threads needed. If this hits the upper bound imposed by the
system, we ignore this though, and just carry on. Sadly the getrlimit()
errnos are not as insightful as I'd wish, because it uses EPERM for a
lot of things other than excessive limits, but this is a good
compromise.
If we fail because of CAP_SYS_RESOURCE, which might remain uncaught
previously in case setrlimit on RLIMIT_NOFILE lowers both limits, it
will remain unreported here. However, I wouldn't want to spam the command
line with such an error message when it's only a very high preset limit
by a user triggering a kernel-limit-overflow.
In case it is a lack of CAP_SYS_RESOURCE, this is later reported when
pthread_create() fails, so we cover this case as well and thus give
the user proper feedback on what he can do.

Signed-off-by: Laslo Hunhold <dev@frign.de>