Ignore queries and fragments in URIs
Previously, a request for "/index.html" would yield a 200, while a
request for "/index.html?foo=bar" would yield a 404, as quark would
look for the file "index.html?foo=bar" in the serve directory.

To accomodate this behaviour, it's no longer sufficient to just compare
realuri and req->uri. Instead, we set a "dirty" flag every time we
change the URI in such a way that it requires a redirect.

According to RFC 3986 section 3, queries and fragments are there
to (further) "identify a resource within the scope of the URI's scheme
and naming authority (if any)". However, it's perfectly legitimate to
just ignore this further specification when the URI itself is already
pointing at a unique resource (i.e. "/index.html").

This behaviour is consistent with dynamic web applications which usually
ignore parameters they don't care about. Quark is too much Zen to care
about any parameters. This has the added bonus that you can now clone
repositories (read-only) via the "dumb" HTTP git-protocol, so

	git clone https://example.org/git/project.git

is now possible (provided you run update-server-info during the
post-update-hook). This wouldn't work previously because git, when
asked to clone via HTTP, would first probe the server with a request for

	project.git/info/refs?service=git-upload-pack

(i.e. asking for the "smart" HTTP git-protocol to confirm). Quark would
return a 404, though, while git only gracefully "downgrades" to the
"dumb" HTTP git-protocol if the request succeeds but only yields a basic
200 response without special git-headers.

This way, it is now trivial to also share git-repositories (and other
gracefully-downgrading protocols). While the "dumb" HTTP git-protocol
only supports read-only-access, I don't think that's much of an overall
loss (to the contrary!).

HTTP authentication is broken and it makes much more sense to enable
ssh-access to contributors and make them push changes via ssh. The key
advantage of HTTP-cloning over git://-cloning is the fact that the git
protocol can be tampered with, while the HTTP-protocol can be encapsulated
into a secure TLS connection.

Signed-off-by: Laslo Hunhold <dev@frign.de>