Fix handling of unexpected hangups
During slowloris-stress-testing I noticed that closing the tool would
pretty reliably trigger a case where quark was continuously getting
edge-triggered on EPOLLIN and notified that there was something to read
on the now definitely closed client-sockets.

This was wrong and due to the fact that I mishandled the case when
read() returns 0 in http_recv_header(). The condition read()==0 and
EPOLL or EWOULDBLOCK does not indicate that you should try again. It's
an EOF-condition and should be handled as such.

Given the request is incomplete at this point, this is handled as a
bad request (HTTP status 400) mostly for the logs at this point, as
being able to transmit the error page itself is highly unlikely.

Signed-off-by: Laslo Hunhold <dev@frign.de>