Refactor queue-event-handling to be pointer-centric
This is one aspect where we can see that kqueue(2) is superior to
epoll(7). While kqueue(2) allows you to both store a pointer and read
out the fd of an event in the ready-list, epoll(7) only provides you
with one of the two (by only offering a union-type-field where you
either store an fd or a pointer).

Previously, wrapper functions would allow to extract either from any
event, which was only non-hacky in the OpenBSD-case, because nothing
stopped you from reading the "fd" from the union that actually was
assigned as a pointer. The distinction was there in the first place
because we are listening on the main incoming socket and other "client"-
sockets at the same time. The main-socket-events contained the socket-fd
and the client-sockets contained the connection-struct-pointer. Now, we
just set the data-pointer of the main socket to NULL to indicate it,
yielding a much more refined usage with one fewer way to screw things
up.

Signed-off-by: Laslo Hunhold <dev@frign.de>