Make queue-event-error-detection stricter
Everything which is not a pollin or pollout is now considered an error.
This is due to the fact how variable epoll(7) is in regard to possible
events, often depending on your kernel configuration (see for instance
EPOLLPRI and possible future additions).

In this context, we also rename the function to better reflect its
purpose.

Signed-off-by: Laslo Hunhold <dev@frign.de>