Apply (D)DoS-Hardening
Until now, if quark found that in case of an incoming connection it
didn't have any vacant connection slots left, it would just not
accept() and thus block any further new connections until a slot was
free.

This may sound reasonable at first, but there are cases where the
connected clients are not benevolent and might firmly occupy all slots
with simple request flooding or more advanced attacks like slowloris,
R-U-Dead-Yet or Slow Read, which all boil down to sending or receiving
data really slowly. The latter attacks are very effective and require
very little resources on the attacker's side.

Thus, the only way to keep things going is to bite the bullet and always
accept a connection, even if it means dropping another connection for
it. In this case, an algorithm determines which client has the most
connections and drops the least-advanced connection of it (i.e. a
connection in the earliest possible stage).

Stress-tests with slowhttptest[0] and siege yielded excellent results,
where quark always remained responsive for the normal visitor despite
an active massive DoS-attack using the aforementioned methods.

Side-note: In the spirit of not having any memory allocations in quark
during "runtime", the constant-space algorithm used to determine the
client with the most connections is quadratic in time over the number
of slots. This, however, is not a big deal, given the number of slots
is always relatively small and such an out-of-connection-scenario is
an edge-case.

[0]:https://github.com/shekyan/slowhttptest
[1]:https://github.com/JoeDog/siege

Signed-off-by: Laslo Hunhold <dev@frign.de>