Refactor server- and connection-logic into separate components
The server-part (creating a pool, launching workers, etc.) and methods
concerning the handling of connections (accepting, serving, logging,
etc.) were all in main.c, making it unnecessarily complicated on top of
the fact that there is already a lot to do there.

Now, these parts are moved into server.h/server.c and connection.h/
connection.c respectively, while main() now just only has to call
into server_init_thread_pool(). In other words, main() now only has to
deal with the argument parsing and preparations, and the rest is
done by other program parts.

The methods are prefixed with server_* and connection_* respectively,
making it much easier to see where each method is defined in the
program.

This refactoring also separates concerns well. The server-struct is
now in server.h (and not in util.h, where it was always out of place)
and spacetok() is moved into util.c.

Signed-off-by: Laslo Hunhold <dev@frign.de>