Use one listening socket for all threads
Previously, we employed SO_REUSEPORT and bound as many listening
sockets as we had threads. Inside the kernel, this creates separate
queues for each socket and a hash-based-round-robin distributes incoming
requests evenly among the listening sockets.

As a result, the load was well-balanced among threads, in contrast to
sharing one listening socket (and one shared queue), because when
epoll() registers new incoming connections, it takes the last-activated
thread (FIFO). As a result, only one thread usually gets most of the
load, unless the server is really stressed.

So why change it back? The reason is latency. It's not a coincidence
that when you're at the supermarket you very often see another queue
pass along much faster than you. Even though a congestion (i.e. slow
cashier or customer) is completely random and no cashier-post is
favoured, all people in the congested queue are affected (the chance
is 2/3 to be in a slower queue when there are 3 queues).
A much more efficient approach is to have one shared queue and 3
cashiers. Even when there's a congestion at one cashier-post, the others
will continue processing customers and the overall latency is much
more consistent.
The same applies to the connection-queue: If a thread is really busy
with a request, the waiting connections will be affected without the
possibility of being processed by another idle thread.
Another reason is efficiency: The hash-based-round-robin in the kernel
adds a small overhead that can be avoided this way, it removes
complexity in the code and undoes the inconsistent "hack" where we
just gave each thread the same listening socket when we were working
with a UNIX-domain socket.

This change adds a small "regression": When quark is hammered with a
lot of connections (>15k/s), it is more motivated to drop connections
because epoll has the aforementioned FIFO-behaviour (i.e. it tends to
pass the connection-events to the same thread), which usually accepts
before processing read/write-related events. This in turn exhausts the
thread's connection-pool at some point, but because it handles that
well, we never enter a DoS-state.

This commit is a preparation for upcoming changes.

Signed-off-by: Laslo Hunhold <dev@frign.de>