Improve connection logging
Given connection_log() can be called on "incomplete" connections since
the last changes regarding properly logging dropped connections, the
function sock_get_inaddr_str() had the broken assumption that when
ss_family was neither AF_INET nor AF_INET6 it would be AF_UNIX, leading
to the wrong logging of just-accepted-and-dropped connections as "uds",
even when listening on an IP socket.

As a solution, the switch is amended and yields the expected
in-address-string "-" when the given sockaddr-struct is not filled yet.

Signed-off-by: Laslo Hunhold <dev@frign.de>